decorated analysisService with kafka sender. This way an external object calls a transactional method instead of an internal method which would not trigger a transaction
