+++ song-server/src/main/java/bio/overture/song/server/controller/AnalysisController.java
@@ -16,12 +16,6 @@
  */
 package bio.overture.song.server.controller;
 
-import static bio.overture.song.core.utils.Separators.COMMA;
-import static bio.overture.song.server.repository.search.IdSearchRequest.createIdSearchRequest;
-import static org.springframework.http.HttpHeaders.AUTHORIZATION;
-import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
-import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
-
 import bio.overture.song.server.model.analysis.Analysis;
 import bio.overture.song.server.model.entity.FileEntity;
 import bio.overture.song.server.repository.search.IdSearchRequest;
@@ -31,11 +25,11 @@ import com.google.common.collect.ImmutableSet;
 import io.swagger.annotations.Api;
 import io.swagger.annotations.ApiOperation;
 import io.swagger.annotations.ApiParam;
-import java.util.List;
-import lombok.RequiredArgsConstructor;
+import lombok.NonNull;
 import lombok.SneakyThrows;
 import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.beans.factory.annotation.Qualifier;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.web.bind.annotation.GetMapping;
@@ -49,8 +43,15 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.ResponseBody;
 import org.springframework.web.bind.annotation.RestController;
 
+import java.util.List;
+
+import static org.springframework.http.HttpHeaders.AUTHORIZATION;
+import static org.springframework.http.MediaType.APPLICATION_JSON_UTF8_VALUE;
+import static org.springframework.http.MediaType.APPLICATION_JSON_VALUE;
+import static bio.overture.song.core.utils.Separators.COMMA;
+import static bio.overture.song.server.repository.search.IdSearchRequest.createIdSearchRequest;
+
 @RestController
-@RequiredArgsConstructor
 @RequestMapping("/studies/{studyId}/analysis")
 @Api(
     tags = "Analysis",
@@ -71,7 +72,13 @@ public class AnalysisController {
           + "}";
 
   /** Dependencies */
-  @Autowired private final AnalysisService analysisService;
+  private final AnalysisService analysisService;
+
+  @Autowired
+  public AnalysisController(
+      @NonNull AnalysisService analysisService) {
+    this.analysisService = analysisService;
+  }
 
   @ApiOperation(
       value = "GetAnalysesForStudy",
