+++ song-server/src/main/java/bio/overture/song/server/service/FileModificationService.java
@@ -36,6 +36,7 @@ import bio.overture.song.core.model.enums.FileUpdateTypes;
 import bio.overture.song.server.converter.FileConverter;
 import bio.overture.song.server.model.entity.FileEntity;
 import javax.transaction.Transactional;
+
 import lombok.NonNull;
 import lombok.val;
 import org.springframework.beans.factory.annotation.Autowired;
