+++ song-server/src/main/java/bio/overture/song/server/service/SubmitService.java
@@ -32,7 +32,7 @@ import bio.overture.song.core.model.SubmitResponse;
 import bio.overture.song.server.model.dto.Payload;
 import com.fasterxml.jackson.databind.JsonNode;
 import java.io.IOException;
-import javax.transaction.Transactional;
+
 import lombok.NonNull;
 import lombok.SneakyThrows;
 import lombok.extern.slf4j.Slf4j;
@@ -58,7 +58,6 @@ public class SubmitService {
     this.studyService = studyService;
   }
 
-  @Transactional
   public SubmitResponse submit(@NonNull String studyId, String payloadString) {
     // Check study exists
     studyService.checkStudyExist(studyId);
