+++ song-server/src/test/java/bio/overture/song/server/service/FileServiceTest.java
@@ -38,6 +38,7 @@ import bio.overture.song.core.utils.JsonUtils;
 import bio.overture.song.core.utils.RandomGenerator;
 import bio.overture.song.server.model.entity.FileEntity;
 import javax.transaction.Transactional;
+
 import lombok.extern.slf4j.Slf4j;
 import lombok.val;
 import org.junit.Before;
